centralised the hero text/header
changed the here ids to classes and added a container to the hero text. Added display flex so it could be centered on the page and moved it below the hero image with flex-end. By adding flex grow of 1 to the child text it allows the background colour to grow to the max width of 650px which matches all the other sections